Check`glossaryUrl` for language-selector.tsx.
This commit uses the URL param `glossaryUrl` to look up languages in the language selector component.

The idea is that authors can then easily add the correct glossary dashboard params in their custom reports.

Feature story:
A teacher should only see available translations for a glossary in the language settings of the dashboard.

[#169452886]
https://www.pivotaltracker.com/story/show/169452886