Update Firebase to use named app and to unsubscribe from watchers
The Firebase functions were updated to use a named app instance so that it does not conflict with the Firebase app in the Activity Player.  All snapshot subscriptions are also tracked so that they can be unsubscribed to when switching pages in the activity player (which otherwise would show a console error as the snapshot would fail during signout).

 Also added secure:no-certs start script.