Added UI to support file uploads in model authoring [#181946122]
Note: this is only the UI support.  The actual uploading is not yet implemented.  Uploading will require a change in LARA before the commented out code in this PR can be used.

This also fixes an incorrect placeholder in an unrelated field in the translation form (the digging deeper field).