Fixed multiple file upload field setting [#181946122]
This actually a general fix for both the definition and translation forms.  Before the field change handler was wrapped in a useCallback but this failed when multiple uploads were in flight.  The change is to remove the useCallback and change the translation setter to use the callback form.  This ensures that the current translation object is always updated instead of the one captured by the useCallback function wrapper.